refactor: refactor TestCommandExecutor
