+++ src/test/java/com/github/zxl0714/redismock/executor/TestCommandExecutor.java
@@ -173,43 +173,6 @@ public class TestCommandExecutor {
     }
 
     @Test
-    public void testExpire() throws ParseErrorException, InterruptedException, EOFException {
-        assertCommandEquals(0, array("expire", "ab", "1"));
-        assertCommandOK(array("SET", "ab", "abd"));
-        assertCommandEquals(1, array("expire", "ab", "1"));
-        assertCommandEquals("abd", array("GET", "ab"));
-        assertCommandError(array("expire", "ab", "a"));
-        Thread.sleep(1000);
-        assertCommandNull(array("GET", "ab"));
-    }
-
-    @Test
-    public void testTTL() throws ParseErrorException, InterruptedException, EOFException {
-        assertCommandEquals(-2, array("ttl", "ab"));
-        assertCommandOK(array("SET", "ab", "abd"));
-        assertCommandEquals(-1, array("ttl", "ab"));
-        assertCommandEquals(1, array("expire", "ab", "2"));
-        assertCommandEquals(2, array("ttl", "ab"));
-        Thread.sleep(1000);
-        assertCommandEquals(1, array("ttl", "ab"));
-        Thread.sleep(1000);
-        assertCommandEquals(-2, array("ttl", "ab"));
-    }
-
-    @Test
-    public void testPTTL() throws ParseErrorException, InterruptedException, EOFException, IOException {
-        assertCommandEquals(-2, array("pttl", "ab"));
-        assertCommandOK(array("SET", "ab", "abd"));
-        assertCommandEquals(-1, array("pttl", "ab"));
-        assertCommandEquals(1, array("expire", "ab", "2"));
-        assertTrue(executor.execCommand(parse(array("pttl", "ab")), socket).compareTo(Response.integer(1900L)) > 0);
-        Thread.sleep(1100);
-        assertTrue(executor.execCommand(parse(array("pttl", "ab")), socket).compareTo(Response.integer(999L)) < 0);
-        Thread.sleep(1000);
-        assertCommandEquals(-2, array("pttl", "ab"));
-    }
-
-    @Test
     public void testIncr() throws ParseErrorException, EOFException {
         assertCommandEquals(1, array("incr", "a"));
         assertCommandEquals(2, array("incr", "a"));
@@ -343,54 +306,6 @@ public class TestCommandExecutor {
     }
 
     @Test
-    public void testDel() throws ParseErrorException, EOFException {
-        assertCommandOK(array("set", "a", "v"));
-        assertCommandOK(array("set", "b", "v"));
-        assertCommandEquals(2, array("del", "a", "b", "c"));
-        assertCommandNull(array("get", "a"));
-        assertCommandNull(array("get", "b"));
-    }
-
-    @Test
-    public void testExists() throws ParseErrorException, EOFException {
-        assertCommandOK(array("set", "a", "v"));
-        assertCommandEquals(1, array("exists", "a"));
-        assertCommandEquals(0, array("exists", "b"));
-    }
-
-    @Test
-    public void testExpireAt() throws ParseErrorException, EOFException {
-        assertCommandOK(array("set", "a", "v"));
-        assertCommandEquals(1, array("expireat", "a", "1293840000"));
-        assertCommandEquals(0, array("exists", "a"));
-        assertCommandOK(array("set", "a", "v"));
-        long now = System.currentTimeMillis() / 1000 + 5;
-        assertCommandEquals(1, array("expireat", "a", String.valueOf(now)));
-        assertCommandEquals(5, array("ttl", "a"));
-        assertCommandError(array("expireat", "a", "a"));
-    }
-
-    @Test
-    public void testPexpireAt() throws ParseErrorException, EOFException {
-        assertCommandOK(array("set", "a", "v"));
-        assertCommandEquals(1, array("pexpireat", "a", "1293840000000"));
-        assertCommandEquals(0, array("exists", "a"));
-        assertCommandEquals(0, array("pexpireat", "a", "1293840000000"));
-        assertCommandOK(array("set", "a", "v"));
-        long now = System.currentTimeMillis() + 5000;
-        assertCommandEquals(1, array("pexpireat", "a", String.valueOf(now)));
-        assertCommandEquals(5, array("ttl", "a"));
-        assertCommandError(array("pexpireat", "a", "a"));
-    }
-
-    @Test
-    public void testPexpire() throws ParseErrorException, EOFException {
-        assertCommandOK(array("set", "a", "v"));
-        assertCommandEquals(1, array("pexpire", "a", "1500000"));
-        assertCommandEquals(1500, array("ttl", "a"));
-    }
-
-    @Test
     public void testLpush() throws ParseErrorException, EOFException, IOException {
         assertCommandEquals(1, array("lpush", "mylist", "!"));
         assertCommandEquals(3, array("lpush", "mylist", "world", "hello"));
@@ -473,14 +388,6 @@ public class TestCommandExecutor {
     }
 
     @Test
-    public void testKeys() throws ParseErrorException, EOFException, IOException {
-        assertCommandOK(array("set", "prefix:a", "a"));
-        assertCommandOK(array("set", "prefix:b", "b"));
-        assertEquals(array("prefix:a","prefix:b"),
-                executor.execCommand(parse(array("keys", "prefix:*")), socket).toString());
-    }
-
-    @Test
     public void testSelect() throws ParseErrorException, EOFException {
         assertCommandNull(array("get", "ab"));
         assertCommandOK(array("SET", "ab", "abc"));
@@ -533,626 +440,5 @@ public class TestCommandExecutor {
 
 
     }
-
-    @Test
-    public void testHashSetAndGet() throws ParseErrorException, EOFException, IOException {
-        assertCommandNull(array("hget", "set", "a"));
-        assertCommandEquals(1, array("hset", "set", "a", "a"));
-        assertCommandEquals("a", array("hget", "set", "a"));
-        assertCommandEquals(1, array("hset", "set", "b", "b"));
-        assertCommandEquals(1, array("hset", "set", "c", "c"));
-        assertCommandEquals(0, array("hset", "set", "a", "aa"));
-        assertCommandEquals("aa", array("hget", "set", "a"));
-        assertCommandEquals("b", array("hget", "set", "b"));
-        assertCommandEquals("c", array("hget", "set", "c"));
-
-        // error
-        assertCommandError(array("hset", "set", "a"));
-        assertCommandError(array("hset", "set", "a", "a", "a"));
-        assertCommandOK(array("set", "a", "a"));
-        assertCommandError(array("hset", "a", "a", "a"));
-        assertCommandError(array("hget", "a", "a"));
-        assertCommandError(array("hget", "set"));
-        assertCommandError(array("hget", "set", "a", "a"));
-
-    }
-
-    @Test
-    public void testHmget() throws ParseErrorException, EOFException, IOException {
-        assertCommandEquals(1, array("hset", "set", "b", "b"));
-        assertCommandEquals(1, array("hset", "set", "c", "c"));
-        assertCommandEquals(1, array("hset", "set", "a", "aa"));
-
-        assertEquals(array("aa"), executor.execCommand(
-            parse(array("hmget", "set", "a")), socket
-        ).toString());
-        assertEquals(array("aa", "b", "c"), executor.execCommand(
-            parse(array("hmget", "set", "a", "b", "c")), socket
-        ).toString());
-        assertEquals("*1\r\n$-1\r\n", executor.execCommand(
-            parse(array("hmget", "set", "aaa")), socket
-        ).toString());
-        assertEquals("*3\r\n$-1\r\n$2\r\naa\r\n$2\r\naa\r\n", executor.execCommand(
-            parse(array("hmget", "set", "aaa", "a", "a")), socket
-        ).toString());
-
-        // error
-        assertCommandOK(array("set", "a", "a"));
-        assertCommandError(array("hmget", "a", "a"));
-        assertCommandError(array("hmget", "set"));
-    }
-
-    @Test
-    public void testHashDelAndExists() throws ParseErrorException, EOFException {
-        assertCommandEquals(0, array("hexists", "set", "a"));
-        assertCommandEquals(0, array("hdel", "set", "a"));
-
-        assertCommandEquals(1, array("hset", "set", "a", "a"));
-        assertCommandEquals(1, array("hset", "set", "b", "b"));
-        assertCommandEquals(1, array("hset", "set", "c", "c"));
-        assertCommandEquals(1, array("hexists", "set", "a"));
-        assertCommandEquals(1, array("hexists", "set", "b"));
-        assertCommandEquals(1, array("hexists", "set", "c"));
-
-        assertCommandEquals(1, array("hdel", "set", "a", "a"));
-        assertCommandEquals(2, array("hdel", "set", "a", "a", "b", "c"));
-
-        assertCommandEquals(0, array("hexists", "set", "a"));
-        assertCommandEquals(0, array("hexists", "set", "b"));
-        assertCommandEquals(0, array("hexists", "set", "c"));
-
-        // error
-        assertCommandOK(array("set", "a", "a"));
-        assertCommandError(array("hdel", "set"));
-        assertCommandError(array("hdel", "a", "a"));
-        assertCommandError(array("hexists", "a", "a"));
-        assertCommandError(array("hexists", "set"));
-        assertCommandError(array("hexists", "set", "a", "b"));
-    }
-
-    @Test
-    public void testHgetall() throws ParseErrorException, EOFException, IOException {
-        assertEquals(Response.EMPTY_LIST.toString(), executor.execCommand(
-            parse(array("hgetall", "set")), socket
-        ).toString());
-        assertCommandEquals(1, array("hset", "set", "a", "1"));
-        assertEquals(array("a", "1"), executor.execCommand(
-            parse(array("hgetall", "set")), socket
-        ).toString());
-        assertCommandEquals(1, array("hset", "set", "b", "2"));
-        assertEquals(array("a", "1", "b", "2"), executor.execCommand(
-            parse(array("hgetall", "set")), socket
-        ).toString());
-        assertCommandEquals(1, array("hdel", "set", "a"));
-        assertEquals(array("b", "2"), executor.execCommand(
-            parse(array("hgetall", "set")), socket
-        ).toString());
-
-        // error
-        assertCommandOK(array("set", "a", "a"));
-        assertCommandError(array("hgetall", "a"));
-        assertCommandError(array("hgetall"));
-        assertCommandError(array("hgetall", "set", "b"));
-    }
-
-    @Test
-    public void testHincrby() throws ParseErrorException, EOFException {
-        assertCommandEquals(0, array("hincrby", "abc", "a", "0"));
-        assertCommandEquals(2, array("hincrby", "abc", "a", "2"));
-        assertCommandEquals(7, array("hincrby", "abc", "a", "5"));
-        assertCommandEquals(1, array("hset", "set", "a", "1"));
-        assertCommandEquals(-2, array("hincrby", "set", "a", "-3"));
-        assertCommandEquals(-3, array("hincrby", "set", "b", "-3"));
-
-        // error
-        assertCommandOK(array("set", "a", "a"));
-        assertCommandError(array("hincrby", "a", "a", "-3"));
-        assertCommandError(array("hincrby", "set", "a"));
-        assertCommandError(array("hincrby", "set", "a", "-3", "-a"));
-        // 对字符串操作
-        assertCommandEquals(0, array("hset", "set", "b", "b"));
-        assertCommandError(array("hincrby", "set", "b", "-3"));
-        // 传入字符串
-        assertCommandError(array("hincrby", "set", "a", "abc"));
-        // 输入范围超限
-        assertCommandError(array("hincrby", "set", "s1", "9223372036854775808"));
-        assertCommandError(array("hincrby", "set", "s2", "-9223372036854775809"));
-        // 累加范围超限
-        assertCommandEquals(9223372036854775807L, array("hincrby", "set", "s1", "9223372036854775807"));
-        assertCommandEquals(-9223372036854775808L, array("hincrby", "set", "s2", "-9223372036854775808"));
-        assertCommandError(array("hincrby", "set", "s1", "1"));
-        assertCommandError(array("hincrby", "set", "s2", "-1"));
-
-    }
-
-    @Test
-    public void testHincrbyfloat() throws ParseErrorException, EOFException {
-        assertCommandEquals("1.21999999999999997", array("hincrbyfloat", "hash", "h1", "1.22"));
-        assertCommandEquals("1.22000000012199994", array("hincrbyfloat", "hash", "h1", "1.22E-10"));
-        assertCommandEquals("1219999999999999962334747426816", array("hincrbyfloat", "hash", "h1", "1.22E30"));
-        assertCommandEquals("179769313486231570814527423731704356798070567525844996598917476803157260780028538760589558632766878171540458953514382464234321326889464182768467546703537516986049910576551282076245490090389328944075868508455133942304583236903222948165808559332123348274797826204144723168738177180919299881250404026184124858368",
-            array("hincrbyfloat", "hash", "hh11", "1.7976931348623157E308"));
-
-        //ERROR
-        assertCommandOK(array("set", "s1", "s1"));
-        assertCommandEquals(1, array("hset", "hash", "ha", "h1"));
-        assertCommandError(array("hincrbyfloat", "hash", "ha", "1"));
-        assertCommandError(array("hincrbyfloat", "hash", "h1", "a"));
-        assertCommandError(array("hincrbyfloat", "s1", "s1", "1.22"));
-        assertCommandError(array("hincrbyfloat", "hash", "h1"));
-        assertCommandError(array("hincrbyfloat", "hash", "h1", "1", "1"));
-        assertCommandError(array("hincrbyfloat", "hash", "h2", "1.7976931348623157E309"));
-        assertCommandError(array("hincrbyfloat", "hash", "hh11", "1.7976931348623157E308"));
-    }
-
-    @Test
-    public void testHkeys() throws ParseErrorException, EOFException, IOException {
-        assertEquals(Response.EMPTY_LIST.toString(), executor.execCommand(
-            parse(array("hkeys", "set")), socket
-        ).toString());
-        assertCommandEquals(1, array("hset", "set", "a", "1"));
-        assertEquals(array("a"), executor.execCommand(
-            parse(array("hkeys", "set")), socket
-        ).toString());
-        assertCommandEquals(1, array("hset", "set", "b", "2"));
-        assertEquals(array("a", "b"), executor.execCommand(
-            parse(array("hkeys", "set")), socket
-        ).toString());
-
-        //error
-        assertCommandOK(array("set", "a", "a"));
-        assertCommandError(array("hkeys", "a"));
-        assertCommandError(array("hkeys"));
-        assertCommandError(array("hkeys", "set", "a"));
-    }
-
-    @Test
-    public void testHlen() throws ParseErrorException, EOFException {
-        assertCommandEquals(0, array("hlen", "set"));
-        assertCommandEquals(1, array("hset", "set", "a", "1"));
-        assertCommandEquals(1, array("hlen", "set"));
-        assertCommandEquals(1, array("hset", "set", "b", "2"));
-        assertCommandEquals(1, array("hset", "set", "c", "3"));
-        assertCommandEquals(3, array("hlen", "set"));
-
-        // error
-        assertCommandOK(array("set", "a", "a"));
-        assertCommandError(array("hlen", "a"));
-        assertCommandError(array("hlen"));
-        assertCommandError(array("hlen", "set", "a"));
-    }
-
-    @Test
-    public void testHmset() throws ParseErrorException, EOFException {
-        assertCommandOK(array("hmset", "set", "a", "1", "b", "2"));
-        assertCommandEquals("1", array("hget", "set", "a"));
-        assertCommandEquals("2", array("hget", "set", "b"));
-        assertCommandOK(array("hmset", "set", "a", "a", "b", "b"));
-        assertCommandEquals("a", array("hget", "set", "a"));
-        assertCommandOK(array("hmset", "set", "a", "1", "a", "2"));
-        assertCommandEquals("2", array("hget", "set", "a"));
-
-        // error
-        assertCommandError(array("hmset", "set", "a"));
-        assertCommandError(array("hmset", "set", "a", "1", "b"));
-        assertCommandError(array("hmset", "set"));
-        assertCommandOK(array("set", "a", "a"));
-        assertCommandError(array("hmset", "a", "a", "1"));
-    }
-
-    @Test
-    public void testHsetnx() throws ParseErrorException, EOFException {
-        assertCommandEquals(1, array("hsetnx", "h1", "key", "val"));
-        assertCommandEquals("val", array("hget", "h1", "key"));
-        assertCommandEquals(0, array("hsetnx", "h1", "key", "newval"));
-        assertCommandEquals("val", array("hget", "h1", "key"));
-
-        // error
-        assertCommandOK(array("set", "s1", "str"));
-        assertCommandError(array("hsetnx", "s1", "key", "val"));
-        assertCommandError(array("hsetnx", "h1", "key"));
-        assertCommandError(array("hsetnx", "h1", "key", "val", "val"));
-    }
-
-    @Test
-    public void testHstrlen() throws ParseErrorException, EOFException {
-        assertCommandEquals(0, array("hstrlen", "h1", "key"));
-        assertCommandEquals(1, array("hsetnx", "h1", "key", "val"));
-        assertCommandEquals(3, array("hstrlen", "h1", "key"));
-        assertCommandEquals(0, array("hset", "h1", "key", "helloworld"));
-        assertCommandEquals(10, array("hstrlen", "h1", "key"));
-        assertCommandEquals(1, array("hset", "h1", "key1", "-10"));
-        assertCommandEquals(3, array("hstrlen", "h1", "key1"));
-        assertCommandEquals(0, array("hstrlen", "h1", "keynull"));
-
-        // error
-        assertCommandOK(array("set", "s1", "str"));
-        assertCommandError(array("hstrlen", "s1", "key"));
-        assertCommandError(array("hstrlen", "h1", "key", "key"));
-        assertCommandError(array("hstrlen", "h1"));
-
-    }
-
-    @Test
-    public void testHvals() throws ParseErrorException, EOFException, IOException {
-        assertEquals(Response.EMPTY_LIST.toString(), executor.execCommand(
-            parse(array("hvals", "h1")), socket
-        ).toString());
-        assertCommandOK(array("hmset", "h1", "k1", "v1", "k2", "v2"));
-        assertEquals(array("v1", "v2"), executor.execCommand(
-            parse(array("hvals", "h1")), socket
-        ).toString());
-
-        // error
-        assertCommandOK(array("set", "s1", "str"));
-        assertCommandError(array("hvals", "s1"));
-        assertCommandError(array("hvals", "h1", "h1"));
-        assertCommandError(array("hvals"));
-    }
-
-    @Test
-    public void testHscan() throws ParseErrorException, EOFException, IOException {
-        // normal
-        assertEquals("*2\r\n$1\r\n0\r\n*-1\r\n",
-            exec(array("hscan", "h1", "0")));
-        assertCommandOK(array("hmset", "h1",
-            "k1", "v1",
-            "k2", "v2",
-            "k3", "v3",
-            "k4", "v4",
-            "k5", "v5",
-            "k6", "v6",
-            "k7", "v7",
-            "k8", "v8",
-            "k9", "v9",
-            "k10", "v10"));
-        assertEquals("*2\r\n$1\r\n0\r\n*20\r\n" +
-            "$2\r\nk1\r\n$2\r\nv1\r\n" +
-            "$2\r\nk2\r\n$2\r\nv2\r\n" +
-            "$2\r\nk3\r\n$2\r\nv3\r\n" +
-            "$2\r\nk4\r\n$2\r\nv4\r\n" +
-            "$2\r\nk5\r\n$2\r\nv5\r\n" +
-            "$3\r\nk10\r\n$3\r\nv10\r\n" +
-            "$2\r\nk6\r\n$2\r\nv6\r\n" +
-            "$2\r\nk7\r\n$2\r\nv7\r\n" +
-            "$2\r\nk8\r\n$2\r\nv8\r\n" +
-            "$2\r\nk9\r\n$2\r\nv9\r\n",
-            exec(array("hscan", "h1", "0")));
-        // count
-        assertEquals("*2\r\n$1\r\n5\r\n*10\r\n" +
-            "$2\r\nk1\r\n$2\r\nv1\r\n" +
-            "$2\r\nk2\r\n$2\r\nv2\r\n" +
-            "$2\r\nk3\r\n$2\r\nv3\r\n" +
-            "$2\r\nk4\r\n$2\r\nv4\r\n" +
-            "$2\r\nk5\r\n$2\r\nv5\r\n",
-            exec(array("hscan", "h1", "0", "count", "5")));
-        assertEquals("*2\r\n$1\r\n0\r\n*10\r\n" +
-            "$3\r\nk10\r\n$3\r\nv10\r\n" +
-            "$2\r\nk6\r\n$2\r\nv6\r\n" +
-            "$2\r\nk7\r\n$2\r\nv7\r\n" +
-            "$2\r\nk8\r\n$2\r\nv8\r\n" +
-            "$2\r\nk9\r\n$2\r\nv9\r\n",
-            exec(array("hscan", "h1", "5", "count", "5")));
-        // match
-        assertEquals("*2\r\n$1\r\n5\r\n*2\r\n" +
-                "$2\r\nk1\r\n$2\r\nv1\r\n",
-            exec(array("hscan", "h1", "0", "count", "5", "match", "k1")));
-        assertEquals("*2\r\n$1\r\n0\r\n*4\r\n" +
-                "$2\r\nk1\r\n$2\r\nv1\r\n" +
-                "$3\r\nk10\r\n$3\r\nv10\r\n",
-            exec(array("hscan", "h1", "0", "count", "10", "match", "k1*")));
-        // error
-        assertCommandOK(array("set", "s1", "str"));
-        assertCommandError(array("hscan", "s1", "0"));
-        assertCommandError(array("hscan", "h1"));
-        assertCommandError(array("hscan", "h1", "-1"));
-        assertCommandError(array("hscan", "h1", "abc"));
-        assertCommandError(array("hscan", "h1", "0", "count", "0"));
-        assertCommandError(array("hscan", "h1", "0", "count", "-1"));
-        assertCommandError(array("hscan", "h1", "0", "count", "abc"));
-        assertCommandError(array("hscan", "h1", "0", "count"));
-        assertCommandError(array("hscan", "h1", "0", "baba", "0"));
-        assertCommandError(array("hscan", "h1", "0", "count", "1", "match"));
-        assertCommandError(array("hscan", "h1", "0", "count", "1", "match", "*", "a"));
-    }
-
-    @Test
-    public void testSet() throws ParseErrorException, EOFException, IOException {
-        // SADD SCARD SMEMBERS
-        assertCommandEquals(0, array("scard", "set1"));
-        assertEquals(array(), exec(array("smembers", "set1")));
-        assertCommandEquals(1, array("sadd", "set1", "k1"));
-        assertEquals(array("k1"), exec(array("smembers", "set1")));
-        assertCommandEquals(1, array("scard", "set1"));
-        assertCommandEquals(0, array("sadd", "set1", "k1"));
-        assertCommandEquals(1, array("scard", "set1"));
-        assertCommandEquals(3, array("sadd", "set1", "k2", "k3", "k4"));
-        assertEquals(array("k2", "k3", "k4", "k1"), exec(array("smembers", "set1")));
-        assertCommandEquals(2, array("sadd", "set1", "k4", "k5", "k6"));
-        assertCommandEquals(6, array("scard", "set1"));
-        // SDIFF
-        assertCommandEquals(3, array("sadd", "set2", "k2", "k3", "k4"));
-        assertCommandEquals(3, array("sadd", "set3", "k4", "k5", "k7"));
-        assertEquals(array("k5", "k6", "k1"), exec(array("sdiff", "set1", "set2")));
-        assertEquals(array("k2", "k3", "k6", "k1"), exec(array("sdiff", "set1", "set3")));
-        assertEquals(array("k6", "k1"), exec(array("sdiff", "set1", "set2", "set3")));
-        assertEquals(array(), exec(array("sdiff", "set2", "set1")));
-        assertEquals(array("k2", "k3", "k4"), exec(array("sdiff", "set2")));
-        assertEquals(array("k2", "k3", "k4"), exec(array("sdiff", "set2", "set0")));
-        // SDIFFSTORE
-        assertCommandEquals(3, array("sdiffstore", "set4", "set2"));
-        assertEquals(array("k2", "k3", "k4"), exec(array("smembers", "set4")));
-        assertCommandEquals(3, array("sdiffstore", "set4", "set1", "set2"));
-        assertEquals(array("k6", "k5", "k1"), exec(array("smembers", "set4")));
-        assertCommandEquals(0, array("sdiffstore", "set4", "set2", "set1"));
-        assertEquals(array(), exec(array("smembers", "set4")));
-        assertCommandOK(array("set", "str1", "str1"));
-        assertCommandEquals(3, array("sdiffstore", "str1", "set1", "set2"));
-        assertEquals(array("k6", "k5", "k1"), exec(array("smembers", "str1")));
-        /*
-        * SINTER
-        * set1: k1 k2 k3 k4 k5 k6
-        * set2: k2 k3 k4
-        * set3: k4 k5 k7
-        * */
-        assertEquals(array("k4", "k5"), exec(array("sinter", "set1", "set3")));
-        assertEquals(array("k4"), exec(array("sinter", "set1", "set2", "set3")));
-        assertEquals(array("k2", "k3", "k4"), exec(array("sinter", "set2")));
-        assertEquals(array(), exec(array("sinter", "set2", "set0")));
-        // SINTERSTORE
-        assertCommandEquals(2, array("sinterstore", "inter_set1", "set1", "set3"));
-        assertEquals(array("k4", "k5"), exec(array("smembers", "inter_set1")));
-        assertCommandEquals(1, array("sinterstore", "inter_set1", "set1", "set3", "set2"));
-        assertEquals(array("k4"), exec(array("smembers", "inter_set1")));
-        assertCommandEquals(0, array("sinterstore", "inter_set1", "set1", "set0"));
-        assertEquals(array(), exec(array("smembers", "inter_set1")));
-        assertCommandEquals(3, array("sinterstore", "inter_set1", "set2"));
-        assertEquals(array("k2", "k3", "k4"), exec(array("smembers", "inter_set1")));
-        // SISMEMBER SMOVE
-        assertCommandEquals(1, array("sismember", "set1", "k1"));
-        assertCommandEquals(1, array("sismember", "set1", "k2"));
-        assertCommandEquals(0, array("sismember", "set1", "k10"));
-        assertCommandEquals(0, array("smove", "set1", "move_set1", "k10"));
-        assertCommandEquals(0, array("sismember", "move_set1", "k10"));
-        assertCommandEquals(1, array("sadd", "set1", "k10"));
-        assertCommandEquals(1, array("smove", "set1", "move_set1", "k10"));
-        assertCommandEquals(1, array("sismember", "move_set1", "k10"));
-        assertCommandEquals(1, array("sadd", "set1", "k10"));
-        assertCommandEquals(1, array("smove", "set1", "move_set1", "k10"));
-        // SPOP SRANDMEMBER
-        assertCommandEquals(4, array("sadd", "set5", "k1", "k2", "k3", "k4"));
-        exec(array("srandmember", "set5"));
-        exec(array("srandmember", "set5", "2"));
-        exec(array("srandmember", "set5", "10"));
-        exec(array("srandmember", "set5", "-2"));
-        exec(array("srandmember", "set5", "0"));
-        assertCommandEquals(4, array("scard", "set5"));
-        exec(array("spop", "set5"));
-        assertCommandEquals(3, array("scard", "set5"));
-        exec(array("spop", "set5", "10"));
-        assertCommandNull(array("spop", "set5"));
-        assertCommandNull(array("srandmember", "set5"));
-        assertCommandEquals(0, array("scard", "set5"));
-        // SREM
-        assertCommandEquals(4, array("sadd", "set5", "k1", "k2", "k3", "k4"));
-        assertCommandEquals(0, array("srem", "set5", "k10"));
-        assertCommandEquals(1, array("srem", "set5", "k1"));
-        assertCommandEquals(2, array("srem", "set5", "k2", "k3", "k1"));
-        // SSCAN
-        assertEquals("*2\r\n$1\r\n0\r\n*-1\r\n",
-            exec(array("sscan", "set6", "0")));
-        assertCommandEquals(5, array("sadd", "set6", "k1", "k2", "k3", "k4", "k5"));
-        assertEquals("*2\r\n$1\r\n3\r\n*3\r\n" +
-                "$2\r\nk2\r\n$2\r\nk3\r\n" +
-                "$2\r\nk4\r\n",
-            exec(array("sscan", "set6", "0", "count", "3")));
-        assertEquals("*2\r\n$1\r\n0\r\n*2\r\n" +
-                "$2\r\nk5\r\n$2\r\nk1\r\n",
-            exec(array("sscan", "set6", "3", "count", "3")));
-        assertEquals("*2\r\n$1\r\n0\r\n*1\r\n" +
-                "$2\r\nk1\r\n",
-            exec(array("sscan", "set6", "0", "count", "5", "match", "k1")));
-        assertEquals("*2\r\n$1\r\n0\r\n*1\r\n" +
-                "$2\r\nk1\r\n",
-            exec(array("sscan", "set6", "0", "count", "5", "match", "*1")));
-        // SUNION SUNIONSTORE
-        assertEquals(array("k2", "k3", "k4"), exec(array("sunion", "set2")));
-        assertEquals(array("k2", "k3", "k4"), exec(array("sunion", "set2", "set0")));
-        assertEquals(array("k2", "k3", "k4", "k5", "k7"), exec(array("sunion", "set2", "set3")));
-        assertEquals(array("k1", "k2", "k3", "k4", "k5", "k6", "k7"), exec(array("sunion", "set1", "set2", "set3")));
-        assertEquals(array(), exec(array("sunion", "set0", "set00", "set000")));
-        assertCommandEquals(0, array("sunionstore", "union_set1", "set0", "set00"));
-        assertCommandEquals(5, array("sunionstore", "union_set1", "set2", "set3"));
-        assertCommandEquals(7, array("sunionstore", "union_set1", "set2", "set3", "set1"));
-        // error
-        assertCommandOK(array("set", "str1", "str1"));
-        assertCommandError(array("sadd", "str1", "k1"));
-        assertCommandError(array("sadd", "set1"));
-        assertCommandError(array("smembers", "str1"));
-        assertCommandError(array("smembers"));
-        assertCommandError(array("smembers", "set1", "set2"));
-        assertCommandError(array("scard", "str1"));
-        assertCommandError(array("scard"));
-        assertCommandError(array("scard", "set1", "set2"));
-        assertCommandError(array("sdiff", "str1", "set2"));
-        assertCommandError(array("sdiff", "set1", "str1"));
-        assertCommandError(array("sdiff", "set1", "set2", "str1"));
-        assertCommandError(array("sdiffstore", "set4"));
-        assertCommandError(array("sdiffstore", "set4", "set1", "str1"));
-        assertCommandError(array("sinter"));
-        assertCommandError(array("sinter", "str1", "set2"));
-        assertCommandError(array("sinter", "set2", "str1"));
-        assertCommandError(array("sinter", "set2", "set1", "str1"));
-        assertCommandError(array("sinterstore", "inter_set1"));
-        assertCommandError(array("sinterstore", "inter_set1", "str1"));
-        assertCommandError(array("sismember", "str1", "k1"));
-        assertCommandError(array("sismember"));
-        assertCommandError(array("sismember", "set1", "k1", "k2"));
-        assertCommandError(array("smove", "set1", "set2"));
-        assertCommandError(array("smove", "set1", "str1", "k1"));
-        assertCommandError(array("smove", "set1", "set2", "k1", "k2"));
-        assertCommandError(array("srandmember", "str1"));
-        assertCommandError(array("srandmember"));
-        assertCommandError(array("srandmember", "set5", "aaa"));
-        assertCommandError(array("srandmember", "set5", "5", "11"));
-        assertCommandError(array("spop", "set5", "-1"));
-        assertCommandError(array("spop", "set5", "a"));
-        assertCommandError(array("spop", "set5", "1", "1"));
-        assertCommandError(array("spop", "str1"));
-        assertCommandError(array("srem", "set5"));
-        assertCommandError(array("srem", "str1", "v1"));
-        assertCommandError(array("sunion", "str1", "set1"));
-        assertCommandError(array("sunion", "set1", "str1"));
-        assertCommandError(array("sunion"));
-        assertCommandError(array("sunionstore", "union_set1", "str1"));
-        assertCommandError(array("sunionstore", "union_set1"));
-        assertCommandError(array("sscan", "str1", "0"));
-        assertCommandError(array("sscan", "set1"));
-        assertCommandError(array("sscan", "set1", "-1"));
-        assertCommandError(array("sscan", "set1", "abc"));
-        assertCommandError(array("sscan", "set1", "0", "count", "0"));
-        assertCommandError(array("sscan", "set1", "0", "count", "-1"));
-        assertCommandError(array("sscan", "set1", "0", "count", "abc"));
-        assertCommandError(array("sscan", "set1", "0", "count"));
-        assertCommandError(array("sscan", "set1", "0", "baba", "0"));
-        assertCommandError(array("sscan", "set1", "0", "count", "1", "match"));
-        assertCommandError(array("sscan", "set1", "0", "count", "1", "match", "*", "a"));
-    }
-
-    @Test
-    public void testMove() throws ParseErrorException, EOFException {
-        assertCommandOK(array("select", "0"));
-        assertCommandEquals(0, array("move", "str1", "1"));
-        assertCommandOK(array("set", "str1", "s1"));
-        assertCommandOK(array("set", "str2", "s2"));
-        assertCommandEquals(1, array("move", "str1", "1"));
-        assertCommandNull(array("get", "str1"));
-        assertCommandOK(array("select", "1"));
-        assertCommandEquals("s1", array("get", "str1"));
-        assertCommandOK(array("set", "str2", "ss2"));
-        assertCommandEquals(0, array("move", "str12", "0"));
-        assertCommandOK(array("select", "0"));
-        assertCommandEquals("s2", array("get", "str2"));
-        //error
-        assertCommandError(array("move", "str1", "abc"));
-        assertCommandError(array("move", "str1", "20"));
-        assertCommandError(array("move", "str1", "-10"));
-        assertCommandError(array("move", "str1", "0"));
-        assertCommandError(array("move", "str1"));
-        assertCommandError(array("move", "str1", "1", "1"));
-    }
-
-    @Test
-    public void testPersist() throws ParseErrorException, EOFException {
-        assertCommandOK(array("set", "k1", "k1"));
-        assertCommandEquals(1, array("expire", "k1", "5"));
-        assertCommandEquals(5, array("ttl", "k1"));
-        assertCommandEquals(1, array("persist", "k1"));
-        assertCommandEquals(-1, array("ttl", "k1"));
-        assertCommandEquals(0, array("persist", "k1"));
-        // error
-        assertCommandError(array("persist"));
-        assertCommandError(array("persist", "k1", "k2"));
-    }
-
-    @Test
-    public void testRandomkey() throws ParseErrorException, EOFException {
-        assertCommandNull(array("randomkey"));
-        assertCommandOK(array("set", "k1", "k1"));
-        assertCommandEquals("k1", array("randomkey"));
-        // error
-        assertCommandError(array("randomkey", "aaa"));
-    }
-
-    @Test
-    public void testRename() throws ParseErrorException, EOFException {
-        assertCommandOK(array("set", "k1", "k1"));
-        assertCommandOK(array("rename", "k1", "k2"));
-        assertCommandEquals("k1", array("get", "k2"));
-        assertCommandNull(array("get", "k1"));
-        assertCommandOK(array("set", "k1", "newk"));
-        assertCommandOK(array("rename", "k1", "k2"));
-        assertCommandEquals("newk", array("get", "k2"));
-        // error
-        assertCommandError(array("rename", "k3", "k2"));
-        assertCommandError(array("rename", "k3", "k2", "k3"));
-        assertCommandError(array("rename", "k3"));
-    }
-
-    @Test
-    public void testRenamenx() throws ParseErrorException, EOFException {
-        assertCommandOK(array("set", "k1", "k1"));
-        assertCommandEquals(1, array("renamenx", "k1", "k2"));
-        assertCommandOK(array("set", "k1", "nk"));
-        assertCommandEquals(0, array("renamenx", "k1", "k2"));
-        assertCommandEquals("nk", array("get", "k1"));
-        assertCommandEquals("k1", array("get", "k2"));
-        // error
-        assertCommandError(array("renamenx", "k3", "k2"));
-        assertCommandError(array("renamenx", "k3", "k2", "k3"));
-        assertCommandError(array("renamenx", "k3"));
-    }
-
-    @Test
-    public void testType() throws ParseErrorException, EOFException {
-        assertCommandOK(array("set", "s1", "s1"));
-        assertCommandEquals(1, array("lpush", "l1", "l1"));
-        assertCommandEquals(1, array("sadd", "set1", "set1"));
-        assertCommandEquals(1, array("hset", "h1", "h1", "h1"));
-        // TODO : (snowmeow:2021/8/27) zset,stream
-        assertCommandNONE(array("type", "s0"));
-        assertCommandEquals("string", array("type", "s1"));
-        assertCommandEquals("list", array("type", "l1"));
-        assertCommandEquals("set", array("type", "set1"));
-        assertCommandEquals("hash", array("type", "h1"));
-        // error
-        assertCommandError(array("type"));
-        assertCommandError(array("type", "a", "b"));
-    }
-
-    @Test
-    public void testScan() throws ParseErrorException, EOFException, IOException {
-        assertCommandOK(array("set", "s1", "s1"));
-        assertCommandOK(array("set", "s2", "s2"));
-        assertCommandOK(array("set", "s3", "s3"));
-        assertCommandOK(array("set", "s4", "s4"));
-        assertCommandOK(array("set", "s5", "s5"));
-        assertCommandOK(array("set", "s11", "s11"));
-
-        assertEquals("*2\r\n$1\r\n0\r\n*6\r\n" +
-            "$2\r\ns2\r\n$3\r\ns11\r\n" +
-            "$2\r\ns3\r\n$2\r\ns4\r\n" +
-            "$2\r\ns5\r\n$2\r\ns1\r\n",
-            exec(array("scan", "0")));
-        assertEquals("*2\r\n$1\r\n3\r\n*3\r\n" +
-                "$2\r\ns2\r\n$3\r\ns11\r\n" +
-                "$2\r\ns3\r\n",
-            exec(array("scan", "0", "count", "3")));
-        assertEquals("*2\r\n$1\r\n0\r\n*3\r\n" +
-                "$2\r\ns4\r\n$2\r\ns5\r\n" +
-                "$2\r\ns1\r\n",
-            exec(array("scan", "3", "count", "3")));
-        assertEquals("*2\r\n$1\r\n3\r\n*3\r\n" +
-                "$2\r\ns2\r\n$3\r\ns11\r\n" +
-                "$2\r\ns3\r\n",
-            exec(array("scan", "0", "count", "3", "match", "s*")));
-        assertEquals("*2\r\n$1\r\n3\r\n*1\r\n" +
-                "$3\r\ns11\r\n",
-            exec(array("scan", "0", "count", "3", "match", "s1*")));
-        assertEquals("*2\r\n$1\r\n0\r\n*1\r\n" +
-                "$2\r\ns1\r\n",
-            exec(array("scan", "3", "count", "3", "match", "s1*")));
-
-        // error
-        assertCommandError(array("scan", "-1"));
-        assertCommandError(array("scan", "0", "count", "0"));
-        assertCommandError(array("scan", "0", "count", "-1"));
-        assertCommandError(array("scan", "0", "count", "abc"));
-        assertCommandError(array("scan", "0", "count"));
-        assertCommandError(array("scan", "0", "baba", "3"));
-        assertCommandError(array("scan", "0", "count", "3", "match"));
-        assertCommandError(array("scan", "0", "count", "3", "match", "*", "abc"));
-    }
-
 
 }
